backend/pyproject.toml
[project]
name = "feina-backend"
version = "0.1.0"
description = "feinaLive后端服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "httpx>=0.28.0",
    "brotli>=1.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "blivedm>=0.1.0",
    "bilibili-api-python>=6.0.0",
    "aiohttp>=3.9.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "cryptography>=42.0.0",
    "langgraph>=0.2.0",
    "numpy",
    "scipy",
    "protobuf",
    "pynput",
    "opencv-python",
    "Pillow",
    "pyopengl",
    "onnx",
    "onnxruntime-gpu",
    "tqdm",
    "torch",
    "OneEuroFilter",
    "sounddevice",
    "pyvirtualcam>=0.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]
tts-edge = [
    "edge-tts>=6.1.0",
]
easyvtuber = [
    "mediapipe==0.10.11",
    "pyanime4k==3.1.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
